# shared-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VibeCare API
  version: 0.7.0
  description: Canonical mobile API. FITRUS credentials never cross this boundary.
servers:
  - url: https://api.example.invalid
    description: Replace per environment
security:
  - bearerAuth: []
paths:
...
  /v1/auth/pin:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantCode, pin]
              properties:
                participantCode: { type: string }
                pin: { type: string, pattern: '^\d{6}$' }
      responses:
        '200': { description: Access and refresh tokens }
        '401': { description: Invalid credentials }
        '423': { description: Locked for 15 minutes after five failures }
  /v1/auth/refresh:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string, minLength: 20 }
      responses:
        '200':
          description: New 15-minute access token; the refresh token itself is not rotated
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, expiresInSec]
                properties:
                  accessToken: { type: string }
                  expiresInSec: { type: integer, const: 900 }
        '400': { description: Invalid JSON, missing token, or authentication secret unavailable }
        '401': { description: Invalid or revoked refresh token }
...
  /v1/feedback-adjustment:
    get:
      description: Returns the participant's current conservative feedback adjustment, or null when no completed-session feedback exists.
      responses:
        '200':
          description: Current feedback adjustment
          content:
            application/json:
              schema:
                type: object
                required: [adjustment]
                properties:
                  adjustment:
                    oneOf:
                      - type: 'null'
                      - type: object
                        required: [intensityCap, requiresReview, reason, reasonCode, policyVersion, sourceSessionId]
                        properties:
                          intensityCap: { type: number, minimum: 0, maximum: 100 }
                          requiresReview: { type: boolean }
                          reason: { type: string }
                          reasonCode: { type: string }
                          policyVersion: { type: string }
                          sourceSessionId: { type: string }
        '401': { description: Missing or invalid access token }
  /v1/session-feedback:
    post:
      description: Idempotent by participant and session; conflicting retries rejected. policyVersion feedback-0.2.0 and reasonCode accompany the adjustment. Physical telemetry is nullable and participant-reported, not device-attested.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId, rpe, pain, dizziness]
              properties:
                sessionId: { type: string }
                rpe: { type: integer, minimum: 0, maximum: 10 }
                pain: { type: integer, minimum: 0, maximum: 10 }
                dizziness: { type: boolean }
                intensityRating: { type: string, enum: [weak, suitable, strong] }
                durationRating: { type: string, enum: [weak, suitable, strong] }
                frequencyRating: { type: string, enum: [weak, suitable, strong] }
                discomfort:
                  type: [string, 'null']
                  maxLength: 500
                  description: Optional participant-entered discomfort note
                earlyStopped: { type: boolean }
                actualDurationSec: { type: [number, 'null'], minimum: 0 }
                measuredPeakG: { type: [number, 'null'], minimum: 0 }
                measuredRmsG: { type: [number, 'null'], minimum: 0 }
      responses:
        '200': { description: Feedback stored }
        '400': { description: Invalid feedback values }
        '401': { description: Missing or invalid access token }
        '404': { description: Session not found for the authenticated participant }
        '409': { description: Conflicting feedback or non-terminal session }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...